Bunch of changes

+++ update-repo.sh
@@ -2,11 +2,35 @@
 
 # Copy dotfiles from home dir to repo
 
-cp ~/.vimrc .
-cp ~/.tmux.conf .
-cp ~/.bashrc .
-cp ~/.zshrc .
-cp ~/.gitconfig .
-cp ~/.localrc .
-cp ~/.ideavimrc .
-cp ~/.config/lazygit/config.yml lazygit_config.yml
+if [[ -f ~/.vimrc ]]; then 
+  cp ~/.vimrc .
+fi
+
+if [[ -f ~/.tmux.conf ]]; then 
+  cp ~/.tmux.conf .
+fi
+
+if [[ -f ~/.bashrc ]]; then 
+  cp ~/.bashrc .
+fi
+
+if [[ -f ~/.zshrc ]]; then 
+  cp ~/.zshrc .
+fi
+
+if [[ -f ~/.gitconfig ]]; then 
+  cp ~/.gitconfig .
+fi
+
+if [[ -f ~/.localrc ]]; then 
+  cp ~/.localrc .
+fi
+
+if [[ -f ~/.ideavimrc ]]; then 
+  cp ~/.ideavimrc .
+fi
+
+if [[ -f ~/.config/lazygit/config.yml ]]; then 
+  cp ~/.config/lazygit/config.yml lazygit_config.yml
+fi
+
